Add the logic and testing with Api Rest, gRPC Local

+++ euro/src/main/java/aposta/register/euro/entities/CreditData.java
@@ -2,15 +2,15 @@ package aposta.register.euro.entities;
 public class CreditData {
     
     private Long creditAAccountId;
-    private Float ammount;
+    private Integer ammount;
 
     public CreditData()
     {
         this.creditAAccountId = 0L;
-        this.ammount = 0.0f;
+        this.ammount = 0;
     }
 
-    public CreditData(Long creditAAccountId, Float ammount) 
+    public CreditData(Long creditAAccountId, Integer ammount) 
     {
         this.setCreditAAccountId(creditAAccountId);
         this.setAmmount(ammount);
@@ -32,11 +32,11 @@ public class CreditData {
        
     }
 
-    public Float getAmmount() {
+    public Integer getAmmount() {
         return ammount;
     }
 
-    public void setAmmount(Float ammount) {
+    public void setAmmount(Integer ammount) {
         if ((ammount == null) || (ammount<=0.0))
         {
              System.out.println("Inválid Amount! The amomount must be more than 0.0");
